etl/postgres_to_es/requirements.txt
psycopg2-binary >= 2.9.10
elasticsearch <9.0.0
redis >= 6.2.0
backoff >= 1.11.1
pydantic>=1.10.0
pydantic-settings >= 2.9.0